+++ backend/vpnathome/apps/openvpn/serializers.py
@@ -1,8 +1,7 @@
+from datetime import datetime
 import uuid
-import ipaddress
 from django.urls import reverse
 from django.utils.text import slugify
-
 from rest_framework import serializers
 
 from vpnathome.apps.x509.models import Ca, Cert
@@ -53,6 +52,19 @@ class AdminServerSerializer(ServerSerializer):
         return reverse('openvpn:download-server-config', kwargs=kwargs)
 
 
+def validity_end_from_time(now, count, unit):
+    from dateutil.relativedelta import relativedelta
+    t = now
+    t.replace(hour=0, minute=0, second=0, microsecond=0)
+    if unit == 'days':
+        t += relativedelta(days=count)
+    elif unit == 'months':
+        t += relativedelta(months=count)
+    elif unit == 'years':
+        t += relativedelta(years=count)
+    return t
+
+
 class CreateServerSerializer(serializers.Serializer):
 
     name = serializers.CharField(max_length=models.Server.MAX_NAME_LENGTH)
@@ -62,15 +74,21 @@ class CreateServerSerializer(serializers.Serializer):
     protocol = serializers.RegexField(regex="tcp|udp", required=False, default="udp")
     network = NetworkAddressSerializerField(required=False, default=models.Server.DEFAULT_NETWORK)
     deploy_dns = serializers.BooleanField(required=False, default=False)
+    validity_time = serializers.IntegerField(required=False, default=1, min_value=1, max_value=1000)
+    validity_time_unit = serializers.ChoiceField(required=False, default="years", choices=["days", "months", "years"])
 
     def create(self, validated_data):
         owner = self.context['owner']
         ca_name = '{name} CA'.format(name=validated_data['name'])
 
+        validity_end = validity_end_from_time(now=datetime.now(),
+                                              count=validated_data['validity_time'],
+                                              unit=validated_data['validity_time_unit'])
         ca = Ca.objects.create(owner=owner,
                                name=ca_name,
                                email=owner.email,
-                               common_name=slugify(ca_name))
+                               common_name=slugify(ca_name),
+                               validity_end=validity_end)
 
         cert_name = '{name} Server Certificate'.format(name=validated_data['name'])
         common_name = uuid.uuid4().hex
@@ -79,7 +97,8 @@ class CreateServerSerializer(serializers.Serializer):
                                    name=cert_name,
                                    type=Cert.TYPE_SERVER,
                                    email=owner.email,
-                                   common_name=common_name)
+                                   common_name=common_name,
+                                   validity_end=validity_end)
 
         tls_auth_key = generate_tls_auth_key()
         dhparams = self.context['dhparams']
@@ -126,6 +145,8 @@ class ClientSerializer(serializers.ModelSerializer):
 class CreateClientSerializer(serializers.Serializer):
 
     name = serializers.CharField(max_length=models.Client.MAX_NAME_LENGTH)
+    validity_time = serializers.IntegerField(required=False, default=1, min_value=1, max_value=1000)
+    validity_time_unit = serializers.ChoiceField(required=False, default="years", choices=["days", "months", "years"])
 
     def create(self, validated_data):
         owner = self.context['owner']
@@ -133,12 +154,17 @@ class CreateClientSerializer(serializers.Serializer):
         cert_name = '{name} Client Certificate'.format(name=validated_data['name'])
         common_name = uuid.uuid4().hex
 
+        validity_end = validity_end_from_time(now=datetime.now(),
+                                              count=validated_data['validity_time'],
+                                              unit=validated_data['validity_time_unit'])
+
         cert = Cert.objects.create(owner=owner,
                                    ca=server.ca,
                                    name=cert_name,
                                    type=Cert.TYPE_CLIENT,
                                    email=owner.email,
-                                   common_name=common_name)
+                                   common_name=common_name,
+                                   validity_end=validity_end)
 
         client = models.Client.objects.create(name=validated_data['name'],
                                               owner=owner,